{{cookiecutter.project_slug}}/requirements/base.txt
fastapi==0.42.0
requests==2.22.0
uvicorn==0.9.1
python-dotenv==0.10.3
pytest==5.2.1
pytest-cov==2.8.1